backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
colorthief
transformers
torch
timm